Avoid refactoring Xtext resources with empty contents
There was an NPE when attempting to refactor an FB type with empty
contents. This can happen for SIFBs or CFBs, which do not have an ST
source, as well as SFBs and BFBs with no algorithms or methods. This
fixes the problem by not attempting to refactor such types with Xtext
in the first place.