Do not update package name for simple renames in ST refactoring
The package name was also updated for simple renames in the ST
refactoring participant. This could cause a mismatch between the package
name seen by Xtext and the actual package name, in cases where the
package name did not conform to the folder structure. This fixes the
problem by not updating the package name for simple renames.