LibraryManager didn't Correctly Handle Closing Projects
When closing project the library manager tried to resolve the
dependencies of the closed project which resulted in a broken
typelibrary instance. On project reopening this led to broken types.

Furthermore as part of this work we could remove a workspace job from
the resouce change listener making it cleaner and hopefully more stable.