Fix missing output in process console for deployment
The deployment process was using the job state to determine whether it
was terminated. This also meant that it considered itself terminated
for a short time before the job was scheduled during launch. If the
corresponding process console was initialized during that time, it then
prematurely closed its streams, leading to no output in the console.

This uses a separate flag to determine the terminated state, which is
set only after the job is done.