re-add E_DEMUX
Can you please keep E_DEMUX. These block is in the standard and users expect them to be there.

Update event output comments for demultiplexed outputs based on K value

Added comments to specify the demultiplexed output based on the K value for each event output in the typelib file.