Rework watch error handling to avoid race condition in monitoring
There was a potential race condition when the variable update count was
incremeneted but the watches had not been updated and a previous UI
update was still going through the watches. This could cause false
positives when determining if a watch had errors, leading to a blinking
effect.

This changes the watch error handling to explicitly update the error
state and not rely on the variable update count. It also introduces a
separate error message for inconsitent values on subapp interfaces.