Avoid editor reload when type is first loaded
The type entry no longer has to load the type to provide the type
editable for an open editor. This means when the type is later loaded,
it could trigger a false reload in the editor. This changes the type
entry adapter to check if the type was actually reloaded.