Only set navigation location of type entry adapter on part activation
when the type entry adapter needs to update typed subapps to have the
correct entry it needs to restore the viewer. This lead to wrong editors
getting activated. This change postpones this to the activation of the
part.

As part of this change also the reload checks for bulk editor
synchronisation are moved to the same infrastructure.