Fix editor reload during refactoring
When an editor reload happens during refactoring, a race condition
could occur when performing the reload asynchronously in the UI thread.

This ensures that editor reloads happen synchronously when already in
the UI thread. It also fixes a superfluous reload dialog when the editor
was still dirty when saving the type entry.