Generating Test Infrastructure within IEC 61499
The plug-in generates a test infrastructure directly within IEC 61499
for a test specification. Due to changes with the standard libraries,
this feature did not work anymore at all - the adapter / FB types were
not found, causing NPEs. This fix updates the names to ensure that they
can be identified correctly.