FB Network Interpreter: Link to network was lost
The generated traces of networks are built from the output events that
are generated by each type runtime, as the type contains the
implementation. When storing these events, the link to the respective
instance in the network was lost, so the trace visualization did not
display the events correctly anymore.