Fix lookup of unqualified data types in importer
The importer resolved data types using DataTypeLibrary.getType(),
which immediately returns an error data type when the type could not
be found. That meant that no further lookup using the current package
or imports was attempted. This fixes the lookup to use getTypeIfExists()
and only afterwards fall back to creating error types with getType().
It also fixes handling of null type names, which should use ANY
instead of null.