Reworked SetPositionCommand to correctly handle screen resolution #1978
The position change commands didn't correctly accommodate the the
requested position. It used the requests moveDelta and not the
constraint object provided by GEF Classic that is already correctly
translated for the target object.

As part of that change a first clean-up of the SetPositionCommand was
performed.

Addresses: https://github.com/eclipse-4diac/4diac-ide/issues/1978